feat: Port x-algorithm to buildable OSS stack
- Build System:
    - Created Cargo workspace with members: `home-mixer`, `thunder`, `candidate-pipeline`, `proto`.
    - Added `Cargo.toml` for all crates with standard dependencies (`tokio`, `tonic`, `serde`, etc.).
    - Resolved `tonic` feature conflicts (enabled `gzip`, disabled `zstd`).

- Code Refactoring:
    - `home-mixer`: Replaced `xai_*` internal macros/utils with `tokio::main` and `env_logger`.
    - `thunder`: Fixed crate imports and added `env_logger`.
    - `clients`: Updated all service clients to use `async_trait` and corrected return types.

- Protocol Buffers:
    - Reverse-engineered `home_mixer.proto` from server implementation.
    - Reconstructed `recsys-proto`, `twitter-context-proto`, and `xai-uas-thrift` definitions.

- Fixes:
    - Fixed retweet deduplication logic in `RepostDeduplicationFilter`.
    - Corrected type mismatches (`i64` vs `u64`) in hydrators and scorers.
    - Implemented missing structs and enums in `visibility-filtering`.

- Testing:
    - Added unit tests for `SelfTweetFilter`.

The project now compiles successfully with `cargo build`.

+++ home-mixer/scorers/weighted_scorer.rs
@@ -72,7 +72,7 @@ impl WeightedScorer {
     fn vqv_weight_eligibility(candidate: &PostCandidate) -> f64 {
         if candidate
             .video_duration_ms
-            .is_some_and(|ms| ms > p::MIN_VIDEO_DURATION_MS)
+            .is_some_and(|ms| ms as i64 > p::MIN_VIDEO_DURATION_MS as i64)
         {
             p::VQV_WEIGHT
         } else {
